+++ core/src/main/scala/com/itv/bucky/AmqpClient.scala
@@ -4,8 +4,8 @@ import java.util.concurrent.TimeUnit
 
 import cats.effect._
 import cats.implicits._
+import com.itv.bucky.consume.{ConsumeAction, DeadLetter, Delivery, PublishCommand}
 import com.rabbitmq.client.{Channel => RabbitChannel, Connection => RabbitConnection}
-
 import com.itv.bucky.decl._
 import com.rabbitmq.client.{ConnectionFactory, ShutdownSignalException}
 import com.typesafe.scalalogging.StrictLogging
