+++ core/src/main/scala/com/itv/bucky/pattern/requeue/RequeueTransformer.scala
@@ -2,14 +2,14 @@ package com.itv.bucky.pattern.requeue
 
 import cats.effect.Sync
 import com.typesafe.scalalogging.StrictLogging
-import com.itv.bucky.{Ack, ConsumeAction, DeadLetter, _}
 
 import scala.concurrent.duration.FiniteDuration
 import scala.language.higherKinds
 import scala.util.Try
-
 import cats.implicits._
 import cats.effect.implicits._
+import com.itv.bucky.{ExchangeName, Handler, Publisher, RequeueHandler}
+import com.itv.bucky.consume.{Ack, ConsumeAction, DeadLetter, Delivery, PublishCommand, RequeueConsumeAction}
 
 case class RequeueTransformer[F[_]](
     requeuePublisher: Publisher[F, PublishCommand],
@@ -17,7 +17,7 @@ case class RequeueTransformer[F[_]](
     requeuePolicy: RequeuePolicy,
     onFailure: RequeueConsumeAction,
     onFailureAction: Delivery => F[Unit]
-                                      )(handler: RequeueHandler[F, Delivery])(implicit F: Sync[F])
+)(handler: RequeueHandler[F, Delivery])(implicit F: Sync[F])
     extends Handler[F, Delivery]
     with StrictLogging {
 
@@ -29,9 +29,7 @@ case class RequeueTransformer[F[_]](
       requeueCount       <- Try(requeueCountAnyRef.toString.toInt).toOption
     } yield requeueCount
 
-  private def buildRequeuePublishCommand(delivery: Delivery,
-                                         remainingAttempts: Int,
-                                         requeueAfter: FiniteDuration): PublishCommand = {
+  private def buildRequeuePublishCommand(delivery: Delivery, remainingAttempts: Int, requeueAfter: FiniteDuration): PublishCommand = {
     val properties = delivery.properties
       .withHeader(requeueCountHeaderName -> remainingAttempts.toString)
       .copy(expiration = Some(requeueAfter.toMillis.toString))
@@ -41,31 +39,29 @@ case class RequeueTransformer[F[_]](
 
   override def apply(delivery: Delivery): F[ConsumeAction] = {
     def perform(action: RequeueConsumeAction): F[ConsumeAction] =
-        action match {
-          case com.itv.bucky.Requeue =>
-            remainingAttempts(delivery) match {
-              case Some(value) if value < 1 =>
+      action match {
+        case com.itv.bucky.consume.Requeue =>
+          remainingAttempts(delivery) match {
+            case Some(value) if value < 1 =>
+              onFailureAction(delivery).map(_ => DeadLetter)
+            case Some(value) =>
+              requeuePublisher(buildRequeuePublishCommand(delivery, value - 1, requeuePolicy.requeueAfter)).map(_ => Ack)
+            case None =>
+              if (requeuePolicy.maximumProcessAttempts <= 1)
                 onFailureAction(delivery).map(_ => DeadLetter)
-              case Some(value) =>
-                requeuePublisher(buildRequeuePublishCommand(delivery, value - 1, requeuePolicy.requeueAfter)).map(_ => Ack)
-              case None =>
-                if (requeuePolicy.maximumProcessAttempts <= 1)
-                  onFailureAction(delivery).map(_ => DeadLetter)
-                else {
-                  val initialRemainingAttempts = requeuePolicy.maximumProcessAttempts - 2
-                  requeuePublisher(
-                    buildRequeuePublishCommand(delivery, initialRemainingAttempts, requeuePolicy.requeueAfter)).map(_ => Ack)
-                }
-            }
-          case other: ConsumeAction => F.point(other)
-        }
+              else {
+                val initialRemainingAttempts = requeuePolicy.maximumProcessAttempts - 2
+                requeuePublisher(buildRequeuePublishCommand(delivery, initialRemainingAttempts, requeuePolicy.requeueAfter)).map(_ => Ack)
+              }
+          }
+        case other: ConsumeAction => F.point(other)
+      }
 
     for {
-      wrapper <- F.delay(handler(delivery))
+      wrapper       <- F.delay(handler(delivery))
       handlerAction <- wrapper
       channelAction <- perform(handlerAction)
-    }
-      yield channelAction
+    } yield channelAction
   }
 
 }
