+++ core/src/main/scala/com/itv/bucky/publish/PublishCommandBuilder.scala
@@ -1,4 +1,7 @@
-package com.itv.bucky
+package com.itv.bucky.publish
+
+import com.itv.bucky.consume.PublishCommand
+import com.itv.bucky.{ExchangeName, PayloadMarshaller, RoutingKey}
 
 trait PublishCommandBuilder[T] {
   def toPublishCommand(t: T): PublishCommand
