--- src/it/resources/qpid-passwd
@@ -1,2 +1,0 @@
-guest:guest
-workorderrouter:workorderrouter
